requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
redis==5.0.1
python-dateutil==2.8.2 